+++ .gitignore
@@ -18,3 +18,4 @@ build/
 
 node_modules/
 backend/.env
+.codegpt/head
