+++ client/src/pages/teachers/TeachingMode.js
@@ -1,9 +1,9 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { db } from '../../firebase'; // Import db from the firebase configuration file
 import { collection, getDocs, doc, setDoc } from 'firebase/firestore';
 import axios from 'axios';
-import './Teaching-Mode.css';
+import './TeachingMode.css';
 import Cropper from 'react-cropper';
 import 'cropperjs/dist/cropper.css';
